CHG: Change the copy setting arma to numpy
For rows, col and cubes we cannot directly use the memory
as there is no guarantee that it is its own memory.

e.g. a steal conversion of

arma::Row<double> out = (
    arma::Row<double>(10, arma::fill::ones) +
    arma::Row<double>(10, arma::fill::ones)
)

will result in an exception as we try to free memory we don't
have access to.

This might have something to do with how armadillo doesn't
use temporaries as the below doesn't suffer the same issue:

arma::Row<double> out = (
    arma::Row<double>(10, arma::fill::ones) +
    arma::Row<double>(10, arma::fill::ones).t()
)