BUG: fix win32 configuration for running tests
With MSVC build chain, the build configuration (e.g. Debug or Release)
is set while running compilation or tests not like make in the
configuration step (initial CMake command).
That's why we have to use a generator expression to set the PYTHONPATH
for running tests (not a standard CMAKE_BUILD_TYPE variable)

NB: $<CONFIGURATION> is deprecated since CMake 3.0. $<CONFIG> is recommended.