CHG: Rewritten memory management for cpp to python
The previous memory management involved manual destruction
of the memory pointed to by the armadillo objects. This was
not robust for objects other than matrices.

The new approach creates a heap allocated
instances and move/copy into that. A pointer to the heap allocated
object is stored in the capsule and the default destructor is used.

The new approach allows all instances to be moved out reliably,
hence the default has become to avoid copying for all objects by
default.

New overloads have been added for all objects, now supporting
`const &`, `&&`, `&` with copy parameter and `*` with copy parameter.