CHG: Rewrite of Numpy to Armadillo conversion
A number of changes have been made:

* copy and strict parameters have becomes ints

* Array's can be stolen when passing copy = -1
When stealing the memory of the array is stolen and replaced with
an array containing a single NaN or the a nullptr is set to the data and
the array dereferenced. The latter approach has to be actived by
defining CARMA_HARD_STEAL.
Note that accessing the array after stealing with CARMA_HARD_STEAL will
result in a segfault.

* Copies are now made using Numpy's C API
This change was needed to prevent internal copies of C-contiguous arrays
going out of scope.

* Armadillo becomes owner of memory after stealing or copy
Ownership of the memory is handed over to armadillo proper after
stealing or copying the array.